+++ core/networking/src/main/kotlin/com/kelsos/mbrc/core/networking/NoDefaultConnectionException.kt
@@ -1,0 +1,5 @@
+package com.kelsos.mbrc.core.networking
+
+import java.io.IOException
+
+class NoDefaultConnectionException : IOException("No default connection is configured")